[VL][Delta] Address review: move Delta funcs to delta/ folder, drop reflection
- move DeltaBitmapAggregator.{cc,h} into cpp/velox/operators/functions/delta/
  so Delta native functions live under their own source folder; update the
  include in RegistrationAllFunctions.cc, the test, and CMakeLists
- drop the reflective serialization-format check in the bitmap-aggregator match
  and match on prettyName only. Delta's only BitmapAggregator construction site
  (DMLWithDeletionVectorsHelper) always uses the Portable format, and a
  non-Portable payload is rejected natively by RoaringBitmapArray::deserialize
  (magic-number check), so the reflection was redundant
