+++ cpp/velox/operators/functions/delta/DeltaBitmapAggregator.h
